activerecord: Fix deprecated behaviour of changed_attributes
Pull request https://github.com/rails/rails/pull/25337 made some
breaking changes to the way changed_attributes works for rails 5.1,
but the PR was only supposed to deprecate using it in after
callbacks.

The problem was that `forget_attribute_assignments` in
changes_internally_applied disconnected mutation_tracker from the
current attributes, so didn't receive any changes made to it in
after callbacks. To preserve the deprecated behaviour, changes are
applied to calculate changed_attributes and after a save nested in
an after callback.